Scraper/requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
pymongo
python-dotenv
